Fix iOS Autolinking
Because `RNVectorIcons.podspec` includes a `resources` line AND `react-native-config.js` includes an `assets` line, each font file is being copied into the iOS project twice when using react native >= 0.60 (autolinking). This commit removes the `resources` line from the podspec, making it so that just a single copy of the assets are added to the project. I chose to remove it from the podspec, not the react-native config because the react-native config also copies assets for Android and I didn't want to break that functionality.