Switch to grunt-webfont for generating Feather TTF
The current rfbezerra/svg-to-ttf method of generating the Feather
webfont fails on fonts that have multiple overlapping paths (see #1309
for details). This is a limitation of using FontCustom, which does not
handle this kind of SVG correctly.

grunt-webfont, on the other hand, is able to handle these SVGs just
fine. While it's unmaintained, it's better to have something
unmaintained and working than questionably maintained and not working
(which seems to be the current state of svg-to-ttf)

I've verified that the generated TTF renders correctly with
https://mathew-kurian.github.io/CharacterMap/, but have not done any
verification that this works beyond that.